[FEAT] Rename of UserProfile page
