+++ src/pages/UserProfile/Favourite/components/RatingStars/index.ts
